Fix references so they account for inheritance correctly
