+++ java/build.gradle
@@ -60,12 +60,14 @@ teavm {
     }
     js {
         moduleType = org.teavm.gradle.api.JSModuleType.ES2015
+        obfuscated = false
+        debugInformation = true
         optimization = org.teavm.gradle.api.OptimizationLevel.AGGRESSIVE
     }
     wasmGC {
         addedToWebApp = true
         modularRuntime = true
-//        obfuscated = false
+        obfuscated = false
         optimization = org.teavm.gradle.api.OptimizationLevel.AGGRESSIVE
     }
 }
