+++ java/src/main/java/mcsrc/ClassIndexVisitor.java
@@ -2,7 +2,6 @@ package mcsrc;
 
 import org.objectweb.asm.*;
 
-// Based on code from Enigma
 public class ClassIndexVisitor extends ClassVisitor {
 	private String name;
 
@@ -18,152 +17,16 @@ public class ClassIndexVisitor extends ClassVisitor {
 
 	@Override
 	public FieldVisitor visitField(int access, String name, String desc, String signature, Object value) {
-		indexField(new Entry.Field(this.name, name, desc));
+		Entry.Field field = new Entry.Field(this.name, name, desc);
+		Indexer.addMemberData(field.owner(), field);
 		return super.visitField(access, name, desc, signature, value);
 	}
 
 	@Override
 	public MethodVisitor visitMethod(int access, String name, String desc, String signature, String[] exceptions) {
-		indexMethod(new Entry.Method(this.name, name, desc));
-		return new IndexReferenceMethodVisitor(api, new Entry.Method(this.name, name, desc));
-	}
-
-	private class IndexReferenceMethodVisitor extends MethodVisitor {
-		private final Entry.Method callerEntry;
-
-		IndexReferenceMethodVisitor(int api, Entry.Method callerEntry) {
-            super(api, null);
-            this.callerEntry = callerEntry;
-		}
-
-		@Override
-		public void visitFieldInsn(int opcode, String owner, String name, String descriptor) {
-			switch (opcode) {
-			case Opcodes.GETSTATIC, Opcodes.PUTSTATIC, Opcodes.GETFIELD, Opcodes.PUTFIELD ->
-					indexFieldReference(callerEntry, new Entry.Field(owner, name, descriptor));
-            }
-
-			super.visitFieldInsn(opcode, owner, name, descriptor);
-		}
-
-		@Override
-		public void visitLdcInsn(Object value) {
-			if (value instanceof Type type && (type.getSort() == Type.OBJECT || type.getSort() == Type.ARRAY)) {
-				if (type.getSort() == Type.ARRAY) {
-					type = type.getElementType();
-				}
-
-				indexClassReference(callerEntry, new Entry.Class(type.getInternalName()));
-			}
-
-			super.visitLdcInsn(value);
-		}
-
-		@Override
-		public void visitTypeInsn(int opcode, String type) {
-			if (opcode == Opcodes.INSTANCEOF || opcode == Opcodes.CHECKCAST) {
-				Type classType = Type.getObjectType(type);
-
-				if (classType.getSort() == Type.ARRAY) {
-					classType = classType.getElementType();
-				}
-
-				indexClassReference(callerEntry, new Entry.Class(classType.getInternalName()));
-			}
-
-			super.visitTypeInsn(opcode, type);
-		}
-
-		@Override
-		public void visitMethodInsn(int opcode, String owner, String name, String descriptor, boolean isInterface) {
-			indexMethodReference(callerEntry, new Entry.Method(owner, name, descriptor));
-			super.visitMethodInsn(opcode, owner, name, descriptor, isInterface);
-		}
-
-		@Override
-		public void visitInvokeDynamicInsn(String name, String descriptor, Handle bootstrapMethodHandle, Object... bootstrapMethodArguments) {
-			if ("java/lang/invoke/LambdaMetafactory".equals(bootstrapMethodHandle.getOwner()) && ("metafactory".equals(bootstrapMethodHandle.getName()) || "altMetafactory".equals(bootstrapMethodHandle.getName()))) {
-				Type samMethodType = (Type) bootstrapMethodArguments[0];
-				Handle implMethod = (Handle) bootstrapMethodArguments[1];
-				Type instantiatedMethodType = (Type) bootstrapMethodArguments[2];
-
-				switch (getHandleEntry(implMethod)) {
-                    case Entry.Field field -> indexFieldReference(callerEntry, field);
-                    case Entry.Method method -> indexMethodReference(callerEntry, method);
-                }
-
-				indexMethodDescriptor(callerEntry, descriptor);
-				indexMethodDescriptor(callerEntry, samMethodType.getDescriptor());
-				indexMethodDescriptor(callerEntry, instantiatedMethodType.getDescriptor());
-			}
-
-			super.visitInvokeDynamicInsn(name, descriptor, bootstrapMethodHandle, bootstrapMethodArguments);
-		}
-
-		private static Entry.Member getHandleEntry(Handle handle) {
-			return switch (handle.getTag()) {
-			case Opcodes.H_GETFIELD, Opcodes.H_GETSTATIC, Opcodes.H_PUTFIELD, Opcodes.H_PUTSTATIC ->
-					new Entry.Field(handle.getOwner(), handle.getName(), handle.getDesc());
-			case Opcodes.H_INVOKEINTERFACE, Opcodes.H_INVOKESPECIAL, Opcodes.H_INVOKESTATIC,
-				Opcodes.H_INVOKEVIRTUAL, Opcodes.H_NEWINVOKESPECIAL ->
-					new Entry.Method(handle.getOwner(), handle.getName(), handle.getDesc());
-			default -> throw new RuntimeException("Invalid handle tag " + handle.getTag());
-			};
-		}
-	}
-
-	public void indexMethod(Entry.Method methodEntry) {
+		Entry.Method methodEntry = new Entry.Method(this.name, name, desc);
 		Indexer.addMemberData(methodEntry.owner(), methodEntry);
-		indexMethodDescriptor(methodEntry, methodEntry.desc());
+		return super.visitMethod(access, name, desc, signature, exceptions);
 	}
 
-	private void indexMethodDescriptor(Entry.Method entry, String descriptor) {
-		for (Type typeDescriptor : Type.getArgumentTypes(descriptor)) {
-			indexMethodType(entry, typeDescriptor);
-		}
-
-		indexMethodType(entry, Type.getReturnType(descriptor));
-	}
-
-	private void indexMethodType(Entry.Method method, Type type) {
-		if (type.getSort() == Type.ARRAY) {
-			indexMethodType(method, type.getElementType());
-			return;
-		}
-
-		if (type.getSort() == Type.OBJECT) {
-			Indexer.addReference(type.getInternalName(), method.reference());
-		}
-	}
-
-	public void indexField(Entry.Field field) {
-		Type type = Type.getType(field.desc());
-
-		Indexer.addMemberData(field.owner(), field);
-
-		if (type.getSort() == Type.ARRAY) {
-			indexField(new Entry.Field(field.owner(), field.name(), type.getElementType().getDescriptor()));
-			return;
-		}
-
-		if (type.getSort() == Type.OBJECT) {
-			Indexer.addReference(type.getInternalName(), field.reference());
-		}
-	}
-
-	public void indexClassReference(Entry.Method callerEntry, Entry.Class referencedEntry) {
-		Indexer.addReference(referencedEntry.name(), callerEntry.reference());
-	}
-
-	public void indexMethodReference(Entry.Method callerEntry, Entry.Method referencedEntry) {
-		Indexer.addReference(referencedEntry.str(), callerEntry.reference());
-
-		if (referencedEntry.name().equals("<init>")) {
-			Indexer.addReference(referencedEntry.owner(), callerEntry.reference());
-		}
-	}
-
-	public void indexFieldReference(Entry.Method callerEntry, Entry.Field referencedEntry) {
-		Indexer.addReference(referencedEntry.str(), callerEntry.reference());
-	}
 }
